Use "UNKNOWN" when no tags are present
On repository forks and similar, often no tags
are present leading to travis failures. This now
logs that there is no tag present for `git describe`
and returns "UNKNOWN". This may cause subsequent
code to fail, but those repositories should simply add a
fake tag if it expects a given format:

```
   git describe || git tag -m "fake" v0.0.0
```